revert for WinGW

+++ src/vm/interpreter.cpp
@@ -1951,7 +1951,7 @@ void StackInterpreter::ProcessTrap(StackInstr* instr)
     
   case CPY_INT_STR_ARY: {
     long index = PopInt();
-    long* value_str = program->GetIntStrings()[index];
+    int* value_str = program->GetIntStrings()[index];
     // copy array
     long* array = (long*)PopInt();
     if(!array) {
